refactor: move update adapter interface to persist module
