+++ casbin/persist/adapters/__init__.py
@@ -14,7 +14,7 @@
 
 from .file_adapter import FileAdapter
 from .filtered_file_adapter import FilteredFileAdapter
-from .update_adapter import UpdateAdapter
+from ..update_adapter import UpdateAdapter
 
 # alias import for backwards compatibility
 FilteredAdapter = FilteredFileAdapter
